suppression de l'affichage de debug et fix d'un type

+++ server/src/modules/objects/BasicShape.hpp
@@ -15,7 +15,7 @@
 
 typedef	struct		s_rayon {
   unsigned int		color;
-  int			r;
+  short			r;
 }			t_rayon;
 
 typedef	struct		s_angle {
